requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
spacy==3.7.2
nltk==3.8.1
transformers==4.36.0
torch==2.1.1
prometheus-client==0.19.0
python-dotenv==1.0.0
alembic==1.13.1
httpx==0.25.2
